+++ CHANGELOG.md
@@ -8,6 +8,14 @@ Format is based on [Keep a Changelog 1.1.0](https://keepachangelog.com/en/1.1.0/
 
 ## [Unreleased]
 
+### Added
+
+- Batch embedding through `Encoder.encode_batch()`. Available for all bindings.
+
+### Changed
+
+- `CrossEncoder.rank()` now batches documents internally making re-raking faster. Affects all bindings.
+
 ## [Python v1.7.0, Flutter v2.5.0, Godot v9.6.0, Kotlin v2.2.0, React Native v2.5.0, Swift v2.3.0] - 2026-07-30
 
 ### Added